Created 7-print_last_digit.c

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -10,7 +10,7 @@
 
 int print_last_digit(int n)
 {
-	int value;
+	long long int value;
 
 	if (n > 0)
 	{
